Adding Graph.cs
